Fix prctl argument type
Although prctl is declared as a varargs function, in actuality it takes
unsigned long arguments (after the first int argument). This patch
correctly truncates its argument to the right width.

Without this change, the 32-bit ARM compiler places the 64-bit slack
value in registers r2 and r3, but prctl expects the first argument to be
in r1. For kernel versions < 4.6, set_sched_policy has not been
correctly setting the timer slack value in 32-bit ARM code.

Test: Run on bullhead device. Verified assembly of libcutils.so

+++ libcutils/sched_policy.cpp
@@ -359,7 +359,7 @@ static void set_timerslack_ns(int tid, unsigned long long slack) {
 
     // TODO: Remove when /proc/<tid>/timerslack_ns interface is backported.
     if ((tid == 0) || (tid == gettid())) {
-        if (prctl(PR_SET_TIMERSLACK, slack) == -1) {
+        if (prctl(PR_SET_TIMERSLACK, static_cast<unsigned long>(slack)) == -1) {
             SLOGE("set_timerslack_ns prctl failed: %s\n", strerror(errno));
         }
     }
